example_project/requirements.txt
django
djangorestframework
django-ninja
uritemplate
pyyaml
requests
